Fix varexists parsing and improve documentation
At the moment, the arguments of the #varexists parser function are
expanded no matter if they are actually parsed as page content or not.
This leads to a situation where variable definitions in the arguments of
the #varexists parser function are done unconditionally instead of only
if the respecitve path is taken.

Analogous to the behavior of ParserFunctions and the #var parser
function this shouldn't be the case here. Furthermore, parsing only
the markup actually needed should boost performance.

Bug: T202679
Change-Id: Ief3e03890f04fa04e4077773ff2332f5acfb51b1