Bump version number
It turns out, that there are extensions depending on the $mVariables attribute,
even though it was marked as @private. Extensions using this functionality
are breaking with changing these attributes to actual private ones.

To make this change clear, the version number should be raised. As this
is shouldn't be an actual change if you followed the instructions to begin
with, this is version 2.4.1, not 2.5.

Change-Id: Ieffa514db3bf82d8870d67e51fdd552f84b4b746