Reduce usage of not needed "static" and "pass by reference"
Both is outdated, not necesarry, makes it harder to test this code in
separation, and just harder to read.

* The individual callbacks for the individual parser functions don't
need to be static. This does not even save memory, because there is a
long-living instance in Parser::mExtVariables anyway.

* Passing a Parser object by reference (with the additional &) is
obscure in modern PHP. This was once necessary in very old PHP, because
objects have not been passed by reference once. This is obsolete for a
long time now.

* Same for the "&get()". The fact this is a reference is not used any
more in this code. onParserClearState is mentioned in the comment, but
doesn't use get() any more.

Change-Id: I1e2da5795fc3a2de566004dbdaa86e0dba9dd918